requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-request==1.5.6
execnet==1.7.1
idna==2.10
importlib-metadata==1.6.0
jsonschema==3.2.0
more-itertools==8.2.0
mysql-connector-python==8.0.20
packaging==20.3
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
